Fix issue #32 by checking for " as well as ; when splitting.
Fixes Set-Cookie (v1) style cookie parsing with regards to quoting.

Note that this also subtly changes the way we deal values we return as
it strips the " when it previously didn't.  These are the logical values
so I believe this is corect but it is a change of behaviour.

This may well affect issue #21 as we'll strip the " on read meaning we
should hopefully do the right thing now.