Fix t/issue32.t on old perl versions
perl < 5.12 had a year2038 bug, so instead we just use the next year as expiry
so that it will pass at least until 2036
and until then people should be on newer versions
and possibly even 64-bit platforms so that this will not be a problem.

This does not adjust the weekday, but apparently nobody checks it.

Fixes #58